\name{hsBands}
\docType{data}
\alias{hsBands}
\title{Homo sapiens GRCh37 chromosome banding}
\description{
  Whole list of GRCh37 human chromosome G-banding, from the UCSC repository.
}
\usage{
  data(hsBands)
}
\format{
  \code{track.table} object with 862 rows and the following columns : "name", "chrom", "strand", "start", "end", "stain".
}
\source{
  University of California, Santa Cruz (genome.ucsc.edu)
}
\keyword{datasets}
