\name{hsGenes}
\docType{data}
\alias{hsGenes}
\title{Subset of Homo sapiens GRCh37 genes}
\description{
  8000 randomly choosen GRCh37 human genes, from the NCBI repository.
}
\usage{
  data(hsGenes)
}
\format{
  \code{data.frame} with 8000 rows and the following columns : "chrom", "start", "end" and "name".
}
\source{
  National Center for Biotechnology Information (ftp.ncbi.nih.gov)
}
\keyword{datasets}
